# Conflicts:
#	test/sync/src/androidTest/kotlin/io/realm/test/shared/SyncedRealmTests.kt

+++ packages/cinterop/src/commonMain/kotlin/io/realm/internal/interop/CoreLogLevel.kt
@@ -31,6 +31,6 @@ expect enum class CoreLogLevel {
     val priority: Int
 
     companion object {
-        fun valueFromPriority(priority: Int): CoreLogLevel
+        fun valueFromPriority(priority: Short): CoreLogLevel
     }
 }
